test: E2E 테스트 코드 수정
