+++ cypress/utils/interceptFetch.ts
@@ -1,0 +1,11 @@
+const interceptFetch = (fixture: string) => {
+  cy.intercept(
+    {
+      method: "GET",
+      url: /^https:\/\/api.themoviedb.org\/3\/movie\/popular*/,
+    },
+    { fixture },
+  );
+};
+
+export default interceptFetch;
